backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
openai==1.3.7
httpx==0.25.2
python-dotenv==1.0.0
Pillow==10.1.0
python-dateutil==2.9.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
python-dateutil==2.9.0
anyio==4.7.0
sniffio==1.3.1
starlette==0.41.3
requests
psycopg2-binary
azure-storage-blob
azure-identity
gunicorn
pyngrok==7.0.0 